pid_exists need to work for processes run by other users, too.

+++ fishweb/expmanage.py
@@ -1,3 +1,4 @@
+import errno
 import os
 import shlex
 import signal
@@ -63,12 +64,32 @@ def start_experiment(expname, timelimit, startfromtrig, stimulus, inifile):
     subprocess.Popen(cmdargs)
 
 
-# http://stackoverflow.com/a/568285
+# http://stackoverflow.com/a/6940314
 def pid_exists(pid):
+    """Check whether pid exists in the current process table.
+    UNIX only.
+    """
+    if pid < 0:
+        return False
+    if pid == 0:
+        # According to "man 2 kill" PID 0 refers to every process
+        # in the process group of the calling process.
+        # On certain systems 0 is a valid PID but we have no way
+        # to know that in a portable fashion.
+        raise ValueError('invalid PID 0')
     try:
         os.kill(pid, 0)
-    except OSError:
-        return False
+    except OSError as err:
+        if err.errno == errno.ESRCH:
+            # ESRCH == No such process
+            return False
+        elif err.errno == errno.EPERM:
+            # EPERM clearly means there's a process to deny access to
+            return True
+        else:
+            # According to "man 2 kill" possible error values are
+            # (EINVAL, EPERM, ESRCH)
+            raise
     else:
         return True
